requirements.txt
streamlit
pandas
numpy
plotly
yfinance
pandas-ta
ta
python-dateutil
statsmodels
scikit-learn